+++ core/terraform/network/outputs.tf
@@ -38,6 +38,11 @@ output "resource_processor_subnet_id" {
   value = azurerm_subnet.resource_processor.id
 }
 
+output "airlock_notification_subnet_id" {
+  value = azurerm_subnet.airlock_notification.id
+}
+
+
 # DNS Zones
 
 output "azure_monitor_dns_zone_id" {