+++ templates/shared_services/airlock_notifier/terraform/providers.tf
@@ -3,14 +3,9 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "=3.57.0"
-    }
-    local = {
-      source  = "hashicorp/local"
-      version = "=2.4.0"
+      version = "=3.113.0"
     }
   }
-
   backend "azurerm" {}
 }